requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.22.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
backoff==2.2.1
banks==2.2.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
durationpy==0.9
fastapi==0.115.7
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
fonttools==4.58.5
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
griffe==1.15.0
grpcio==1.70.0
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==32.0.0
langchain==0.3.17
langchain-classic==1.0.1
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.2
langsmith==0.3.3
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.12
llama-index-cli==0.5.3
llama-index-core==0.14.12
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.13
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.11.7
llama-parse==0.6.54
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.24.1
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.60.2
opencv-python==4.12.0.88
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
ormsgpack==1.12.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.5.1
posthog==3.11.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.3
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.2
pyparsing==3.2.3
pypdf==6.6.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
streamlit==1.41.1
striprtf==0.0.26
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uuid_utils==0.13.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
xxhash==3.6.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
